refactor: extract inner classes from CameraMarkerTracker (731→234 lines)
Extract 8 inner classes into package-private top-level files:
- CameraMarkerConfiguration (abstract base)
- PerspectiveCameraMarkerConfiguration (abstract perspective base)
- StartingCameraMarkerConfiguration, LayoutCameraMarkerConfiguration
- OrthographicCameraMarkerConfiguration (abstract ortho base)
- TopCameraMarkerConfiguration, SideCameraMarkerConfiguration, FrontCameraMarkerConfiguration

Public API unchanged. Constants moved to consuming classes.
Package-private accessors added to tracker for extracted classes.
Characterization test verifies API surface and type hierarchy.

Closes #607
